refactor: Output 컴포넌트 분리
