+++ src/components/My/NotFound/index.jsx
@@ -5,7 +5,7 @@ import * as S from "./index.styled";
 function NotFound() {
   return (
     <S.Container>
-      <h1>주문 목록이 비어있습니다.</h1>
+      <h2>주문 목록이 비어있습니다.</h2>
     </S.Container>
   );
 }
