+++ src/components/My/NotFound/index.styled.js
@@ -1,5 +1,10 @@
 import styled from "@emotion/styled";
 
-const Container = styled.div``;
+const Container = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+`;
 
 export { Container };
